Don't throw away empty lines or comment lines
Throwing away comment lines is undesirable because it makes it
impossible to break up long commit messages into paragraphs; this in
turn breaks a lot of functionality in various Git commands which expect
commit messages to consist of a one-line "subject" followed by
additional summary text, separated by a blank line.

So drop the filtering of blank lines and comment lines too, and instead
pass the "--cleanup=strip" flag to "git commit" so that it can perform
the cleanup for us (stripping comments and leading and trailing blank
lines) just as it normally does when "git commit" is invoked from the
command line.

Signed-off-by: Wincent Colaiuta <win@wincent.com>