Perform commit on BufUnload rather than BufWritePost
When writing a lengthy commit message you might want to save the buffer
occasionally out of habit, but if you do you'll find that a commit
takes place and your half-written commit message got used.

To avoid this, change the BufWritePost autocmd in the :GitCommit
implementation were changed to BufUnload.

This also matches the behaviour of the command line when you have
EDITOR=vim (or any other editor) in your environment and do a "git
commit": the actual commit is only processed when you exit the editor,
not when you save the message.

See:

  http://github.com/motemen/git-vim/issues#issue/7

Signed-off-by: Wincent Colaiuta <win@wincent.com>